move css value input container
