+++ apps/builder/app/builder/features/style-panel/sections/backgrounds/background-size.tsx
@@ -2,7 +2,7 @@ import { Grid, Select, theme } from "@webstudio-is/design-system";
 import { toValue } from "@webstudio-is/css-engine";
 import { styleConfigByName } from "../../shared/configs";
 import { parseCssValue } from "@webstudio-is/css-data";
-import { CssValueInputContainer } from "../../controls/position/css-value-input-container";
+import { CssValueInputContainer } from "../../shared/css-value-input";
 import {
   type StyleValue,
   TupleValue,
