+++ apps/builder/app/builder/features/style-panel/sections/borders/border-property.tsx
@@ -7,7 +7,7 @@ import {
   NestedIconLabel,
   ToggleButton,
 } from "@webstudio-is/design-system";
-import { CssValueInputContainer } from "../../controls/position/css-value-input-container";
+import { CssValueInputContainer } from "../../shared/css-value-input";
 import { styleConfigByName } from "../../shared/configs";
 import { PropertyName } from "../../shared/property-name";
 import { getStyleSource } from "../../shared/style-info";
