+++ apps/builder/app/builder/features/style-panel/sections/outline/outline-offset.tsx
@@ -1,6 +1,6 @@
 import type { StyleProperty, UnitValue } from "@webstudio-is/css-engine";
 import { Grid, theme } from "@webstudio-is/design-system";
-import { CssValueInputContainer } from "../../controls/position/css-value-input-container";
+import { CssValueInputContainer } from "../../shared/css-value-input";
 import type { SectionProps } from "../shared/section";
 import { PropertyName } from "../../shared/property-name";
 import { styleConfigByName } from "../../shared/configs";
