+++ apps/builder/app/builder/features/style-panel/shared/css-value-input/index.ts
@@ -1,1 +1,2 @@
 export { CssValueInput, type IntermediateStyleValue } from "./css-value-input";
+export { CssValueInputContainer } from "./css-value-input-container";
